% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateBinary.R
\name{annotateBinary}
\alias{annotateBinary}
\title{Annotate the input genomic regions with annotation of type "binary".}
\usage{
annotateBinary(input_id, input_chr, input_start, input_end,
                annot_id = NULL, annot_chr, annot_start, annot_end)
}
\arguments{
\item{input_id}{Character vector defining the name of input genomic regions (e.g. gene id)}

\item{input_chr}{Character vector defining the name of the chromosome for input genomic regions (one of chr1, chr2, ..., chrX, chrY or chrM).}

\item{input_start}{Numeric vector specifying the starting position of input genomic regions.}

\item{input_end}{Numeric vector specifying the ending position of input genomic regions.}

\item{annot_id}{Optional (NULL by default): Character vector defining the name or ID of each annotation entry. If NULL, the row numbers (indexes) are used as IDs.}

\item{annot_chr}{Character vector defining the name of the chromosome for annotation entries (one of chr1, chr2, ..., chrX, chrY or chrM).}

\item{annot_start}{Numeric vector specifying the starting position of annotation entries.}

\item{annot_end}{Numeric vector specifying the ending position of annotation entries.}
}
\value{
A data.frame containing the input genomic regions and 2 extra columns which specify:
\enumerate{
  \item The number of annotation entries that overlap with each input region.
  \item The IDs (names) of annotation entries that overlap with each input region (separated by comma).
}
}
\description{
This function annotates the input genomic regions with a given "binary" annotation. A "binary" annotation is simply
a set of genomic regions without any extra attribute (e.g. the set of enhancers/promoters or the set of eQTLs).
}
\examples{

}
