% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateCategorical.R
\name{annotateCategorical}
\alias{annotateCategorical}
\title{Annotate the input genomic regions with annotation of type "categorical".}
\usage{
annotateCategorical(input_id, input_chr, input_start, input_end, annot_chr,
  annot_start, annot_end, annot_category)
}
\arguments{
\item{input_id}{Character vector defining the name of input genomic regions (e.g. gene id)}

\item{input_chr}{Character vector defining the name of the chromosome for input genomic regions (one of chr1, chr2, ..., chrX, chrY or chrM).}

\item{input_start}{Numeric vector specifying the starting position of input genomic regions.}

\item{input_end}{Numeric vector specifying the ending position of input genomic regions.}

\item{annot_chr}{Character vector defining the name of the chromosome for annotation entries (one of chr1, chr2, ..., chrX, chrY or chrM).}

\item{annot_start}{Numeric vector specifying the starting position of annotation entries.}

\item{annot_end}{Numeric vector specifying the ending position of annotation entries.}

\item{annot_category}{Character vector specifying categories for annotation entries.}
}
\value{
A data.frame containing the input genomic regions and 2 extra columns which specify:
\enumerate{
  \item The categories that overlap with each input region (separated by comma).
  \item The percentages of overlaps between each input region and its overlapping categories (separated by comma).
}
}
\description{
This function annotates the input genomic regions with a given "categorical" annotation. A "categorical" annotation specifies
a set of genomic regions and assigns a category to each of them (e.g. chromHMM annotation which assigns a chromatin state to
any segment of the genome).
}
\note{
The percentage of overlap between an input genomic region with a specific category is calculated as the
total portion of the input region that is covered by that category devided by the length of the region.
}
\examples{

}
