% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMultiScore.R
\name{annotateMultiScore}
\alias{annotateMultiScore}
\title{Annotate the input genomic regions with annotation of type "multi-score".}
\usage{
annotateMultiScore(input_id, input_chr, input_start, input_end, annot_chr,
  annot_start, annot_end, annot_sub_id, annot_sub_score)
}
\arguments{
\item{input_id}{Character vector defining the name of input genomic regions (e.g. gene id)}

\item{input_chr}{Character vector defining the name of the chromosome for input genomic regions (one of chr1, chr2, ..., chrX, chrY or chrM).}

\item{input_start}{Numeric vector specifying the starting position of input genomic regions.}

\item{input_end}{Numeric vector specifying the ending position of input genomic regions.}

\item{annot_chr}{Character vector defining the name of the chromosome for annotation entries (one of chr1, chr2, ..., chrX, chrY or chrM).}

\item{annot_start}{Numeric vector specifying the starting position of annotation entries.}

\item{annot_end}{Numeric vector specifying the ending position of annotation entries.}

\item{annot_sub_id}{Character vector defining sets of sub-IDs for annotation entries (separated by comma).}

\item{annot_sub_score}{Character vector defining sets of sub-scores for annotation entries (separated by comma).}
}
\value{
A data.frame containing the input genomic regions and 3 extra columns which specify:
\enumerate{
  \item The sub-IDs of annotation entries that overlap with each input region (separated by comma).
  \item The average score of annotation for each overlapping sub-ID over each input region.
  \item The percentage of overlap between each input region and its overlapping sub-IDs.
}
}
\description{
This function annotates the input genomic regions with a given "multi-score" annotation. A "multi-score" annotation specifies
a set of genomic regions and assigns a set of sub-IDs and their corresponding sub-scores to each of them
(e.g. the annotation of DNAse hypersensitive clusters by ENCODE).
}
\note{
For each input genomic region, the average score of annotation for each sub-ID is measured as the weighted average of scores of
overlapping annotation entries where the percentages of overlaps of annotation entries with the input region are taken
as weights (this is measured for each sub-ID separately).
The percentage of overlap between an input genomic region with a specific annotation sub-ID is calculated as the
total portion of the input region that is covered by that annotation sub-ID divided by the length of the region.
}
\examples{

}
