% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHiCinteractions.R
\name{findHiCinteractions}
\alias{findHiCinteractions}
\title{Find interactions between two sets of genomic regions through Hi-C interactions.}
\usage{
findHiCinteractions(input_id, input_chr, input_start, input_end, hic_f1_id,
  hic_f1_chr, hic_f1_start, hic_f1_end, hic_f2_id, hic_f2_chr,
  hic_f2_start, hic_f2_end, target_id, target_chr, target_start,
  target_end)
}
\arguments{
\item{input_id}{Character vector defining the name of input genomic regions (e.g. gene id)}

\item{input_chr}{Character vector defining the name of the chromosome for input genomic regions (one of chr1, chr2, ..., chrX, chrY or chrM).}

\item{input_start}{Numeric vector specifying the starting position of input genomic regions.}

\item{input_end}{Numeric vector specifying the ending position of input genomic regions.}

\item{hic_f1_id}{Character defining the ID for left fragments of Hi-C interactions.}

\item{hic_f1_chr}{Character vector defining the name of the chromosome for left fragments of Hi-C interactions (one of chr1, chr2, ..., chrX, chrY or chrM).}

\item{hic_f1_start}{Numeric vector specifying the starting position of left fragments of Hi-C interactions.}

\item{hic_f1_end}{Numeric vector specifying the ending position of left fragments of Hi-C interactions.}

\item{hic_f2_id}{Character defining the ID for right fragments of Hi-C interactions.}

\item{hic_f2_chr}{Character vector defining the name of the chromosome for right fragments of Hi-C interactions (one of chr1, chr2, ..., chrX, chrY or chrM).}

\item{hic_f2_start}{Numeric vector specifying the starting position of right fragments of Hi-C interactions.}

\item{hic_f2_end}{Numeric vector specifying the ending position of right fragments of Hi-C interactions.}

\item{target_id}{Character vector defining the name of target genomic regions (e.g. gene id)}

\item{target_chr}{Character vector defining the name of the chromosome for target genomic regions (one of chr1, chr2, ..., chrX, chrY or chrM).}

\item{target_start}{Numeric vector specifying the starting position of target genomic regions.}

\item{target_end}{Numeric vector specifying the ending position of target genomic regions.}
}
\value{
A list of two data.frames:
\itemize{
  \item {
    The first data.frame contains the input genomic regions and 3 extra columns which specify:
    \enumerate{
      \item The number of target genomic regions that interact with each input genomic region through Hi-C interactions.
      \item The IDs of target genomic regions that interact with each input genomic region through Hi-C interactions (separated by comma).
      \item The IDs of Hi-C interactions that connect each input genomic region with target genomic regions (separated by comma).
    }
  }
  \item {
    The second data.frame provides further information about each Hi-C interaction. It specifies the genomic regions of
    the left and right Hi-C fragments. Furthermore, the IDs of input genomic regions and target genomic regions that
    interact through these Hi-C interactions are listed in each row (separated by comma).
  }
}
}
\description{
This function finds interactions between input and target genomic regions based on a given set of Hi-C interactions.
It basically finds the subset of given Hi-C interactions in which one fragment overlaps with input genomic regions
and the other fragment overlaps with target genomic regions.
}
\note{
The ID of a Hi-C interaction is defined by concatenating the left fragment ID and the right fragment ID (separated by dash).
}
\examples{

}
