% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRegions.R
\name{testRegions}
\alias{testRegions}
\title{Apply a statistical test to identify significantly mutated genomic regions in each cancer.}
\usage{
testRegions(ROIs, mutational_catalogs)
}
\arguments{
\item{ROIs}{A data.frame containing input genomic regions. The required columns are: \cr
\itemize{
  \item region_id - Defines the name of the genomic regions (e.g. gene id).
  \item chr - The name of the chromosome (one of chr1, chr2, ..., chrX, chrY or chrM).
  \item start - The starting position of the region in the chromosome.
  \item end - The ending position of the region in the chromosome.
}}

\item{mutational_catalogs}{A list of data.frames. Each data.frame contains the mutational catalogue in a specific cancer
and must have the following columns: \cr
\itemize{
  \item subject_id - Defines the id of the subject for which the mutation is recorded (e.g. ICGC donor-id or sample-id).
  \item chr - The name of the chromosome (one of chr1, chr2, ..., chrX, chrY or chrM).
  \item pos - The position of the mutation in the chromosome.
  \item cancer - The cancer type of the mutational catalogue.
}}
}
\value{
A data.frame containing the input genomic regions, plus the calculated p-values for each genomic region and each cancer.
}
\description{
Given a set of input genomic regions and a list of mutational catalogues in various cancers,
a p-value for each genomic region (in each cancer separately) is calculated which can be used
as a measure of how significantly each genomic region is mutated in each cancer
(see References for more details).
}
\examples{

}
\references{
asdfas
}
